FB-496 corrected makeStock field to BigDecimal type

+++ backend-api/src/test/kotlin/com/rarible/flow/api/controller/OrderApiControllerTest.kt
@@ -24,7 +24,6 @@ import org.springframework.context.annotation.Import
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.web.reactive.server.WebTestClient
 import java.math.BigDecimal
-import java.math.BigInteger
 import java.time.LocalDateTime
 import java.time.ZoneOffset
 
@@ -298,7 +297,7 @@ class OrderApiControllerTest {
                 "FLOW",
                 BigDecimal.TEN
             ),
-            makeStock = BigInteger.TEN,
+            makeStock = BigDecimal.TEN,
             lastUpdatedAt = LocalDateTime.now(ZoneOffset.UTC),
             createdAt = LocalDateTime.now(ZoneOffset.UTC),
             type = OrderType.LIST
@@ -328,7 +327,7 @@ class OrderApiControllerTest {
                 value = BigDecimal.valueOf(100L),
                 tokenId = itemId.tokenId
             ),
-            makeStock = BigInteger.TEN,
+            makeStock = BigDecimal.TEN,
             lastUpdatedAt = LocalDateTime.now(ZoneOffset.UTC),
             createdAt = LocalDateTime.now(ZoneOffset.UTC),
             type = OrderType.LIST
